Starting work on background collapse feature
We need a different edge finding approach to more thoroughly mimic
the original crash kiss artwork. The current approach finds edges
of the foreground/background intersection. In reality these subjects
are have organic, poorly defined edges. The key difficulty is that their
edges tend to not be continuous. A strand of hair could hang over the
subject's forehead, for example, and the hair strand would be marked as
the outer edge. This creates gaps of background space in our smashed
images, which is not how the original manual crash kiss works.

These new edge finding strategies will be introduced:

* A strict comparison of the background to the foreground. In this approach,
  the entire image is scanned for the starting and ending of the foreground.
  Edges will be arrays of shape (R, N, 2) where R is the number of rows in the
  image and N is the max number of edges found in any given row. Each edge comes
  as a pair of start and stop indices.
* A compromise between the original analyze-by-chunks approach, which is
  pretty efficient, and the strict approach. A parameter like --edge-depth
  might be added so that the user controls how deep into the image the edge
  finding algorithm should look.