All sides' backgrounds should always be sampled
For issue #2. We want to make smarter decisions about which sides' backgrounds
are reasonable and which are not based on the median background value for each
side of the image. The `--relative-sides` option should only restrict which
edges we detect as that's the only performance intensive operation. It costs basically
nothing to find the background value of each side.

`Subject` will still only iterate over `Side` objects that are specified in
the `--relative-sides` option,
but `Subject.up.background` will now give a value instead of a `NoneType`
exception.