RGB select works for any combination of RGB or A
We can't create a view of the image array for certain combinations
of RGBA, but we can still create expensive copies and get the behavior
we want from the edge finder. When a combination of RGBA results in a
copy of the image array, a warning is raised so that the user can
understand why things are taking twice as long.