Fixed terrible bug in foreground selection
The foreground must be selected with the following logic:
"A pixel is the FG if any of R, G, or B are NOT EQUAL to the BG value"

Previously, the approach was
"A pixel is the FG if all of R, G, or B are NOT EQUAL to the BG value"

Using `np.any` instead of `np.all` fixes the problem.

This fixes issue #5.