Testing required different packing of bounds data
It's easier for testing if the `get_fg_bounds` function returns
a `namedtuple` with a `max_depth` field. This way we don't have to keep
recalculating this number. It will save us against possible bugs elsewhere
in the code, too. The important thing here is that max_depth has a ceiling,
and `get_fg_bounds` ensures that it doesn't exceed that value.