aborting Cython-based OpenMP approach
Cython does its best to infer which variables should be threadlocal.
Unfortunately, it gets it wrong. When you use inplace operators (-=) on a variable,
it'll mark it as a reduction variable, which prevents you from using it outside of
loops. This should be trivial to get around by using C and OpenMP directly.