First working version of new OpenMP-optimized vers
The new C impl works by maintaining left/right inner/outer cursors
and working from the center of the image outward. It should be
more cache-friendly, and it might be faster despite the fact that
none of this will be vectorized. Certain edge cases (like rows of the image
where there is NO foreground) will be slower, but really diabolical cases
(like tons of mixed foreground/background e.g. in between human hairs)
will be faster. Overall, it's a less complicated, more straightforward
implementation. And, since it's in C, there are more ways to squeeze
better performance out of it.