simplified foreground-finding code
Got rid of the "rgb select" feature which let the user
consider one/more of the three RGB channels. It wasn't useful.
Simplified the background value processing code so that
the background value can be a 24-bit integer (i.e. 0xFFFFFF for white).
This means the user can pass in ANY RGB color into the crash kiss
process and it'll find the fg/bg based on this bg color.